+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.7.0] - 2020-05-07
+
 ### Added
 
 - New struct `cmd::ProcessLinesActions` to expose actions available while reading live output from a command.
@@ -146,6 +148,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial version of Rustwide, extracted from Crater.
 
+[0.7.0]: https://github.com/rust-lang/rustwide/releases/tag/0.7.0
 [0.6.1]: https://github.com/rust-lang/rustwide/releases/tag/0.6.1
 [0.6.0]: https://github.com/rust-lang/rustwide/releases/tag/0.6.0
 [0.5.1]: https://github.com/rust-lang/rustwide/releases/tag/0.5.1